ARM/VExpressPkg: StandaloneMM instance of ExtractGuidedSectionLib
The ExtractGuidedSectionLib implementation in Embedded package
creates HOBs to get Uncompress handler information.

Since StandaloneMM is a HOB consumer and not a HOB producer,
a StandaloneMM instance of ExtractGuidedSectionLib has been
introduced.

Therefore, switch to using the StandaloneMM instance of the
ExtractGuidedSectionLib.

Also, when using Lzma compression for StandaloneMm the size
of the image can be reduced.

The image size reduction can be seen as below:
before:
    FVMAIN_COMPACT [96%Full] 1048576 (0x100000) total,
    1016224 (0xf81a0) used, 32352 (0x7e60) free
after:
    FVMAIN_COMPACT [38%Full] 1048576 (0x100000) total,
    402880 (0x625c0) used, 645696 (0x9da40) free

Signed-off-by: Levi Yun <yeoreum.yun@arm.com>